dashboard/requirements.txt
streamlit>=1.33.0
plotly>=5.20.0
pandas>=2.0.0
google-cloud-bigquery>=3.11.0
db-dtypes>=1.2.0
